requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
transformers
requests
pyttsx3
dotenv